+++ Server/src/main/java/arch/galaxyeclipse/server/data/model/LocationObject.java
@@ -82,27 +82,27 @@ public class LocationObject implements Serializable {
         this.rotationAngle = rotationAngle;
     }
 
-    private int positionX;
+    private float positionX;
 
     @Column(name = "position_x", nullable = false, insertable = true, updatable = true, length = 10, precision = 0)
     @Basic
-    public int getPositionX() {
+    public float getPositionX() {
         return positionX;
     }
 
-    public void setPositionX(int positionX) {
+    public void setPositionX(float positionX) {
         this.positionX = positionX;
     }
 
-    private int positionY;
+    private float positionY;
 
     @Column(name = "position_y", nullable = false, insertable = true, updatable = true, length = 10, precision = 0)
     @Basic
-    public int getPositionY() {
+    public float getPositionY() {
         return positionY;
     }
 
-    public void setPositionY(int positionY) {
+    public void setPositionY(float positionY) {
         this.positionY = positionY;
     }
