+++ sql/create.v8.sql
@@ -25,7 +25,7 @@ create table player_activation_hash (
 -- shipType info
 create table ship_state (-- dynamic state of the shipType
   ship_state_id               integer auto_increment primary key,
-  ship_state_move_speed       integer not null default 0,
+  ship_state_move_speed       float   not null default 0,
   ship_state_rotation_speed   float   not null default 0,
   ship_state_hp               integer not null,
   ship_state_armor_durability integer not null
@@ -39,8 +39,8 @@ create table ship_type (
   ship_type_hp_max                integer     not null,
   ship_type_energy_regen          integer     not null,
   ship_type_hp_regen              integer     not null,
-  ship_type_move_max_speed        integer       not null,
-  ship_type_move_acceleration     integer       not null,
+  ship_type_move_max_speed        integer     not null,
+  ship_type_move_acceleration     integer     not null,
   ship_type_rotation_max_speed    float       not null,
   ship_type_rotation_acceleration float       not null,
   weapon_slots_count              integer     not null,
@@ -48,8 +48,8 @@ create table ship_type (
 );
 create table ship_config (-- static state of the shipType
   ship_config_id                    integer auto_increment primary key,
-  ship_config_move_max_speed        integer   not null,
-  ship_config_move_acceleration     integer   not null,
+  ship_config_move_max_speed        float   not null,
+  ship_config_move_acceleration     float   not null,
   ship_config_rotation_max_speed    float   not null,
   ship_config_rotation_acceleration float   not null,
   ship_config_armor                 integer not null,
@@ -111,8 +111,8 @@ create table weapon (
 );
 create table engine (
   engine_id                   integer auto_increment primary key,
-  move_acceleration_bonus     integer   not null,
-  move_max_speed_bonus        integer   not null,
+  move_acceleration_bonus     float   not null,
+  move_max_speed_bonus        float   not null,
   rotation_acceleration_bonus float   not null,
   rotation_max_speed_bonus    float   not null,
   item_id                     integer not null
@@ -130,8 +130,8 @@ create table inventory_item (
 create table location (
   location_id     integer auto_increment primary key,
   location_name   varchar(32) not null,
-  location_width  integer       not null,
-  location_height integer       not null
+  location_width  integer     not null,
+  location_height integer     not null
 );
 create table location_object (
   location_object_id               integer auto_increment primary key,
@@ -139,8 +139,8 @@ create table location_object (
   location_object_type_id          integer not null,
   object_native_id                 integer not null, -- defines the primary key of table depending on object type
   rotation_angle                   float   not null default 0,
-  position_x                       integer   not null,
-  position_y                       integer   not null,
+  position_x                       float   not null,
+  position_y                       float   not null,
 
   location_id                      integer not null
 );
